Minor changes and fixes
- It should now display errors if the data required cannot be validated.
This is still a bit hacky, but I'll work on this in the coming updates.
- Now properly detects wheter the country is US or not, and displays
values in the imperial/metric format. You can press select to change
between units. This will now be saved so the next time you open the app,
it'll display unit in your desired format.
- Made net stuff print in black, on a black background so that users
don't see this. This was mainly for debugging, so I'd rather not make it
noticable in the final product.
- Minor changes to weather data initialization.
- Marked as a safe homebrew from now on.
- Made the gekihen splash screen display for 2 seconds.